requirements.txt
# Xybernetex AI OpenEngine — Python dependencies
# Install with: pip install -r requirements.txt

# ── Core Engine ────────────────────────────────────────────────────────────────
numpy>=1.26.0
torch>=2.0.0
python-dotenv>=1.0.0

# ── ToolServer ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0

# ── Storage & Queue ───────────────────────────────────────────────────────────
psycopg2-binary>=2.9.0
redis>=5.0.0
boto3>=1.34.0

# ── LLM / JSON helpers ────────────────────────────────────────────────────────
json-repair>=0.30.0

# ── Dev / test (optional) ─────────────────────────────────────────────────────
# pytest>=8.0.0
# mypy>=1.8.0
# ruff>=0.4.0
